requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Django==3.2.16
django-allauth==0.52.0
django-environ==0.9.0
django-markdown==0.8.4
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
